+++ RacingManager/src/RacingManagerLN/SubGestaoCP/Carro/C1Hibrido.java
@@ -3,7 +3,7 @@ package RacingManagerLN.SubGestaoCP.Carro;
 public class C1Hibrido extends C1 {
     private int potenciaEletrica;
 
-    public C1Hibrido(String idCarro, String marca, String modelo, int cilindrada, int potenciaCombustao, float pac, tipoPneu tipoPneu, int downforce, modoMotor modoMotor) {
+    public C1Hibrido(String idCarro, String marca, String modelo, int cilindrada, int potenciaCombustao, float pac, tipoPneu tipoPneu, float downforce, modoMotor modoMotor) {
         super(idCarro, marca, modelo, cilindrada, potenciaCombustao, pac, tipoPneu, downforce, modoMotor);
     }
