+++ RacingManager/src/RacingManagerLN/SubGestaoCP/Carro/GT.java
@@ -2,7 +2,7 @@ package RacingManagerLN.SubGestaoCP.Carro;
 
 public class GT extends Carro {
 
-    public GT(String idCarro, String marca, String modelo, int cilindrada, int potenciaCombustao, float pac, tipoPneu tipoPneu, int downforce, modoMotor modoMotor) {
+    public GT(String idCarro, String marca, String modelo, int cilindrada, int potenciaCombustao, float pac, tipoPneu tipoPneu, float downforce, modoMotor modoMotor) {
         super(idCarro, marca, modelo, cilindrada, potenciaCombustao, pac, tipoPneu, downforce, modoMotor);
     }
