+++ RacingManager/src/RacingManagerLN/SubGestaoCP/Carro/GTHibrido.java
@@ -3,7 +3,7 @@ package RacingManagerLN.SubGestaoCP.Carro;
 public class GTHibrido extends GT {
     private int _potenciaEletrica;
 
-    public GTHibrido(String idCarro, String marca, String modelo, int cilindrada, int potenciaCombustao, float pac, tipoPneu tipoPneu, int downforce, modoMotor modoMotor) {
+    public GTHibrido(String idCarro, String marca, String modelo, int cilindrada, int potenciaCombustao, float pac, tipoPneu tipoPneu, float downforce, modoMotor modoMotor) {
         super(idCarro, marca, modelo, cilindrada, potenciaCombustao, pac, tipoPneu, downforce, modoMotor);
     }
